# mvcc: fully concurrent read