This PR solves "large read blocking write" issue (https://github.com/etcd-io/etcd/issues/10525) in mvcc/backend, as an alternative approach to #9384. The idea is to make ALL reads concurrent, therefore avoiding the need to predict and postpone the large reads. 
For 'full story', please refer to the [etcd: Fully Concurrent Reads Design Proposal](https://docs.google.com/document/d/1V9UuL8BnpZF2xFpHhvE1lO2hWvtdiLml3Ukr-m4jcdI/edit#heading=h.wuzq924gzsh).
# TL;DR
In presence of long (expensive) reads, this PR increases the throughput of other requests (short reads on writes) by about 70%, reduces P99 and P999 latency by about 90%. In other benchmark cases where there are no long reads, performance w/ and w/o this PR is comparable. (see later section for detailed results)

# What does this PR solve
### Concurrency bottleneck in current backend implementation
Any ongoing kvstore read transaction will hold a RWMutex read lock [1], which blocks any modification to backend read buffer. There are two places where an backend write transaction needs to make modifications to read buffer:
 - when a kvstore write transaction ends, write buffer (i.e., the changes made by this kvstore write transaction) needs to be merged into read buffer. [2]
 - when a batch interval ends and changes are actually committed, current read buffer and boltdb Tx needs to be reset / re-create for next batch interval. [3]

Therefore a long read transaction can potentially block write. Note that this mutual blocking is necessary to provide linearizable guarantee, simply removing the block will not work. Backend needs to make sure the combined view of read buffer + bolddb Tx is always consistent.

Ref:
[1] 
https://github.com/etcd-io/etcd/blob/886d30d223a0c97f214ac0902bc150d656cd4101/mvcc/kvstore_txn.go#L32-L34
https://github.com/etcd-io/etcd/blob/886d30d223a0c97f214ac0902bc150d656cd4101/mvcc/kvstore_txn.go#L53-L54

[2] https://github.com/etcd-io/etcd/blob/886d30d223a0c97f214ac0902bc150d656cd4101/mvcc/backend/batch_tx.go#L278-L280

[3] https://github.com/etcd-io/etcd/blob/886d30d223a0c97f214ac0902bc150d656cd4101/mvcc/backend/batch_tx.go#L302-L304

### How this PR addresses the bottleneck
 - In the beginning of a kvstore read transaction, a copy of the current backend read buffer is created. Any range ops in this transaction will read from the copy and boltdb Tx. Therefore this transaction does not need to block following changes to backend read buffer.
 - At the end of a batch interval, we do not rollback boltdb Tx immediately. Instead, we create new boltdb Tx (and reset read buffer) and immediately starts the next batch interval. Any old boltdb Tx will be closed only when all kvstore read transactions that are referencing it are done.

# Benchmark result (PR rebased on 5/8/2019)
### Setup
Generated 100K keys
```
tools/benchmark/benchmark put --key-space-size 100000 --key-size 256 --val-size 10240 --total 100000 --sequential-keys
```

## Throughput benchmark
#### 1. [<u>Short reads</u>](https://github.com/etcd-io/etcd/pull/10523#issuecomment-490752874)
Result is comparable.

#### 2. Writes
Result is comparable.
https://github.com/etcd-io/etcd/pull/10523#issuecomment-490758729

#### 3. Short reads + Writes
Result is comparable, this PR is slight worse in P99 and P999 read latency distribution. 
This is likely to be the worst case scenario for this PR. Since the server is overloaded by writes, the backend read buffer is loaded with pending writes.
https://github.com/etcd-io/etcd/pull/10523#issuecomment-490763739

#### 4. Long reads + Writes
Result: this PR increased write throughput by **69.5%**, reduced P99 write latency by **90.8%**, reduced P999 write latency by **85.3%**.
This is the scenario this PR tries to improve.
https://github.com/etcd-io/etcd/pull/10523#issuecomment-490768752

#### 5. Long reads + (short) Reads + Writes
Result: In presence of long expensive reads, short read throughput is increased by **77.5%**, P99 read latency is reduced by **93.0%**, P999 read latency is reduced by **82.9%**; and write throughput is increased by **85.3%**, P99 write latency is reduced by **94.4%**, P999 write latency is reduced by **83.9%**.
https://github.com/etcd-io/etcd/pull/10523#issuecomment-491419911

## Kubernetes scalability test result 
Posted here: https://github.com/etcd-io/etcd/pull/10523#issuecomment-499262001