@jingyih 2019-03-06
**Benchmark**

Setup:
We used the same setup as in https://github.com/etcd-io/etcd/pull/9384#issuecomment-462566915, where we have 3 types of client requests running concurrently: long read, short read and write.

Result:
**-This PR**
Short read response time histogram:
  0.0006 [1]	|
  0.0741 [2609]	|∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎
  0.1475 [0]	|
  0.2210 [0]	|
  0.2945 [0]	|
  0.3680 [17]	|
  0.4414 [121]	|∎
  0.5149 [151]	|∎∎
  0.5884 [79]	|∎
  0.6618 [12]	|
  0.7353 [10]	|

Write response time histogram:
  0.0004 [1]	|
  0.0736 [638]	|∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎
  0.1468 [57]	|∎∎∎
  0.2199 [55]	|∎∎∎
  0.2931 [57]	|∎∎∎
  0.3663 [56]	|∎∎∎
  0.4395 [75]	|∎∎∎∎
  0.5127 [44]	|∎∎
  0.5859 [12]	|
  0.6590 [3]	|
  0.7322 [2]	|

Long read response time histogram:
  1.7072 [1]	|∎∎∎
  1.7822 [3]	|∎∎∎∎∎∎∎∎∎∎
  1.8572 [9]	|∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎
  1.9322 [12]	|∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎
  2.0073 [8]	|∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎
  2.0823 [2]	|∎∎∎∎∎∎
  2.1573 [2]	|∎∎∎∎∎∎
  2.2323 [0]	|
  2.3073 [0]	|
  2.3824 [2]	|∎∎∎∎∎∎
  2.4574 [1]	|∎∎∎

**-Master**
Short read response time histogram:
  0.0006 [1]	|
  0.0082 [2831]	|∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎
  0.0158 [36]	|
  0.0234 [22]	|
  0.0309 [24]	|
  0.0385 [20]	|
  0.0461 [21]	|
  0.0537 [26]	|
  0.0613 [14]	|
  0.0689 [4]	|
  0.0765 [1]	|

Write response time histogram:
  0.0004 [1]	|
  0.0068 [936]	|∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎
  0.0131 [28]	|∎
  0.0195 [7]	|
  0.0258 [4]	|
  0.0322 [2]	|
  0.0385 [4]	|
  0.0449 [2]	|
  0.0512 [6]	|
  0.0576 [6]	|
  0.0639 [4]	|

Long read response time histogram:
  1.6783 [1]	|∎∎∎∎∎
  1.7156 [5]	|∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎
  1.7528 [5]	|∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎
  1.7901 [8]	|∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎
  1.8274 [7]	|∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎
  1.8647 [5]	|∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎
  1.9019 [2]	|∎∎∎∎∎∎∎∎∎∎
  1.9392 [2]	|∎∎∎∎∎∎∎∎∎∎
  1.9765 [1]	|∎∎∎∎∎
  2.0138 [2]	|∎∎∎∎∎∎∎∎∎∎
  2.0510 [2]	|∎∎∎∎∎∎∎∎∎∎