@jingyih 2019-03-06
Please ignore the conflict, this PR needs a rebase to #10506.