@xiang90 2019-03-11
@jingyih Have you done a throughput benchmark? Copying the cache might affect throughput but not the latency.