@jingyih 2019-03-11
@xiang90 

BTW, in the previous benchmark, we are already hitting the throughput limit, i.e., the rate we are requesting as described in https://github.com/etcd-io/etcd/pull/9384#issuecomment-462566915 is larger than the actual rate.