@jingyih 2019-05-09
Rebased to master.