@jingyih 2019-05-09
### Throughput comparison: short reads
```
tools/benchmark/benchmark range "A" "Z" --total 300000 --rate 10000 --conns 10 --clients 10
```
(Note that the requested ops per second is higher than actual served ops per second.)
#### The result is comparable. Maybe slightly better with this PR.

#### Master
Summary:
  Total:	45.4235 secs.
  Slowest:	0.0167 secs.
  Fastest:	0.0002 secs.
  Average:	0.0010 secs.
  Stddev:	0.0014 secs.
  Requests/sec:	6604.5057

Response time histogram:
  0.0002 [1]	|
  0.0019 [265438]	|∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎
  0.0035 [17638]	|∎∎
  0.0052 [7170]	|∎
  0.0068 [4330]	|
  0.0084 [3031]	|
  0.0101 [1845]	|
  0.0117 [450]	|
  0.0134 [64]	|
  0.0150 [23]	|
  0.0167 [10]	|

Latency distribution:
  10% in 0.0003 secs.
  25% in 0.0004 secs.
  50% in 0.0005 secs.
  75% in 0.0008 secs.
  90% in 0.0021 secs.
  95% in 0.0039 secs.
  99% in 0.0081 secs.
  99.9% in 0.0105 secs.

#### Master + PR
Summary:
  Total:	44.9010 secs.
  Slowest:	0.0147 secs.
  Fastest:	0.0002 secs.
  Average:	0.0010 secs.
  Stddev:	0.0014 secs.
  Requests/sec:	6681.3650

Response time histogram:
  0.0002 [1]	|
  0.0017 [260698]	|∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎
  0.0031 [20088]	|∎∎∎
  0.0046 [8230]	|∎
  0.0060 [4482]	|
  0.0075 [2960]	|
  0.0089 [2165]	|
  0.0103 [1202]	|
  0.0118 [151]	|
  0.0132 [20]	|
  0.0147 [3]	|

Latency distribution:
  10% in 0.0003 secs.
  25% in 0.0004 secs.
  50% in 0.0005 secs.
  75% in 0.0008 secs.
  90% in 0.0022 secs.
  95% in 0.0037 secs.
  99% in 0.0078 secs.
  99.9% in 0.0101 secs.