@jingyih 2019-05-09
### Throughput comparison: writes
```
tools/benchmark/benchmark put --key-space-size 100 --key-size 256 --val-size 10240 --total 60000 --rate 2000 --clients 10 --conns 10
```
(Note that the requested ops per second is higher than actual served ops per second.)

#### Master
Summary:
  Total:	38.4857 secs.
  Slowest:	0.2721 secs.
  Fastest:	0.0003 secs.
  Average:	0.0041 secs.
  Stddev:	0.0058 secs.
  Requests/sec:	1559.0213

Response time histogram:
  0.0003 [1]	|
  0.0275 [59633]	|∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎
  0.0547 [356]	|
  0.0818 [0]	|
  0.1090 [0]	|
  0.1362 [0]	|
  0.1634 [0]	|
  0.1906 [0]	|
  0.2177 [0]	|
  0.2449 [0]	|
  0.2721 [10]	|

Latency distribution:
  10% in 0.0011 secs.
  25% in 0.0018 secs.
  50% in 0.0030 secs.
  75% in 0.0042 secs.
  90% in 0.0063 secs.
  95% in 0.0153 secs.
  99% in 0.0260 secs.
  99.9% in 0.0333 secs.

#### Master + PR
Summary:
  Total:	38.0085 secs.
  Slowest:	0.0395 secs.
  Fastest:	0.0003 secs.
  Average:	0.0040 secs.
  Stddev:	0.0047 secs.
  Requests/sec:	1578.5939

Response time histogram:
  0.0003 [1]	|
  0.0042 [45641]	|∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎
  0.0081 [10040]	|∎∎∎∎∎∎∎∎
  0.0120 [1093]	|
  0.0160 [333]	|
  0.0199 [822]	|
  0.0238 [1011]	|
  0.0277 [700]	|
  0.0316 [267]	|
  0.0356 [72]	|
  0.0395 [20]	|

Latency distribution:
  10% in 0.0011 secs.
  25% in 0.0018 secs.
  50% in 0.0030 secs.
  75% in 0.0042 secs.
  90% in 0.0059 secs.
  95% in 0.0151 secs.
  99% in 0.0261 secs.
  99.9% in 0.0328 secs.