@jingyih 2019-06-05
## GCE 5k node scalability test result
Sig-scalability lead @wojtek-t helped me running the test.

 - The PR that overwrote etcd image, which is later reverted after few test runs: https://github.com/kubernetes/test-infra/pull/12710
 - Link to the test with new etcd image, which is 1130745634945503235: https://prow.k8s.io/view/gcs/kubernetes-jenkins/logs/ci-kubernetes-e2e-gce-scale-performance/1130745634945503235

### Result compared to baseline
Using test run 1130383116527996934 as baseline. It is right before 1130745634945503235.
 - Expensive LIST requests themselves are comparable. Using "LIST pods" as example. 

| Latency (ms)  | Baseline | Image built from this PR |
| ------------- | ------------- | ------------- |
| P50  | 1927 | 2185 |
| P90  | 5280 | 5349 |
| P99  | 7725 | 7871 |

Ref: http://perf-dash.k8s.io/#/?jobname=gce-5000Nodes&metriccategoryname=APIServer&metricname=DensityResponsiveness&Resource=pods&Scope=cluster&Subresource=&Verb=LIST

 - At the very beginning of the test where there are a lot of expensive "LIST pods", POST requests have much better latency. The example here is "POST clusterrolebindings".

| Latency (ms)  | Baseline | Image built from this PR |
| ------------- | ------------- | ------------- |
| P50  | 2.40 | 2.67 |
| P90  | 67.4 | 11.2 |
| P99  | 993 | 25.4 |

Ref: http://perf-dash.k8s.io/#/?jobname=gce-5000Nodes&metriccategoryname=APIServer&metricname=DensityResponsiveness&Resource=clusterrolebindings&Scope=cluster&Subresource=&Verb=POST

   - "POST pods" also has better latency, this type of request spread through most part of the test.

| Latency (ms)  | Baseline | Image built from this PR |
| ------------- | ------------- | ------------- |
| P50  | 18.4 | 3.16 |
| P90  | 400 | 38.6 |
| P99  | 962 | 218 |

Ref: http://perf-dash.k8s.io/#/?jobname=gce-5000Nodes&metriccategoryname=APIServer&metricname=DensityResponsiveness&Resource=pods&Scope=namespace&Subresource=&Verb=POST